Finpuss

+++ kalkulator_Del4/kalkulator Del4/Syntax.h
@@ -1,11 +1,12 @@
 #pragma once
 #include <iostream>
 #include <string>
+#include <sstream>
 class Syntax
 {
 protected:
 	bool contains(std::string &str, char c);
-	unsigned short int count(std::string &str, char c);
+	int count(std::string &str, char c);
 	void removeSpaces(std::string &str);
 	void replace(std::string &str1, std::string str2, std::string str3);
 	void fixSyntax(std::string &str);
